Move riaksvc package into trafficvault/backends
