+++ traffic_ops/traffic_ops_golang/trafficvault/backends/riaksvc/dsutil.go
